requirements.txt
# Auto Caption Agent - Python Dependencies
# Install with: pip install -r requirements.txt

# OpenAI Whisper: speech-to-text with word/segment timestamps
openai-whisper

# Python bindings for ffmpeg (audio extraction)
ffmpeg-python

# NumPy is needed by Whisper internally
numpy

# Streamlit: web UI for Phase 4
streamlit
